requirements.txt
Flask
Pillow
numpy
tensorflow
